+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add a Toxic and Toxics type for the Go client
 * Add `Dockerfile`
 * Fix latency toxic limiting bandwidth #67
+* Add Slicer toxic
 
 # 1.1.0